+++ doc/CHANGELOG.md
@@ -1,3 +1,25 @@
+<a name="1.1.1"></a>
+## [1.1.1](https://github.com/aurelia/router/compare/1.1.0...v1.1.1) (2017-01-06)
+
+
+### Bug Fixes
+
+* **Router:** fix an error when updateTitle is called but there is not current instruction ([872fda1](https://github.com/aurelia/router/commit/872fda1))
+
+
+
+<a name="1.1.0"></a>
+# [1.1.0](https://github.com/aurelia/router/compare/1.0.7...v1.1.0) (2016-12-03)
+
+
+### Features
+
+* **router:** add extension point to transform titles when the document title is built ([629da8c](https://github.com/aurelia/router/commit/629da8c))
+* **router:** add extension point to transform titles when the document title is built ([2c40ce1](https://github.com/aurelia/router/commit/2c40ce1)), closes [#343](https://github.com/aurelia/router/issues/343)
+* **router:** add extension point to transform titles when the document title is built ([3d42256](https://github.com/aurelia/router/commit/3d42256)), closes [#343](https://github.com/aurelia/router/issues/343)
+
+
+
 <a name="1.0.6"></a>
 ## [1.0.6](https://github.com/aurelia/router/compare/1.0.5...v1.0.6) (2016-10-06)